+++ R/utils-list.R
@@ -47,3 +47,9 @@ list_replace <- function(x, y) {
   }
   x
 }
+
+list_drop_empty <- function(x) {
+  are_empty <- map_vec(x, \(.x) is_empty(.x))
+  x[are_empty] <- NULL
+  x
+}