+++ R/yaml.R
@@ -1,17 +1,30 @@
-schemas_are_up_to_date <- function(current, new) {
-  current <- current[c("author", "affiliations")]
-  identical(current, new)
+eol <- function() {
+  if (.Platform$OS.type == "unix") "\n" else "\r\n" # nocov
 }
 
-json_update <- function(x, json) {
-  x <- yaml.load(x)
-  if (is.null(x)) {
-    return(json)
-  }
-  if (schemas_are_up_to_date(x, json)) {
+as_verbatim_lgl <- function(x) {
+  x <- if_else(x, "true", "false")
+  structure(x, class = "verbatim")
+}
+
+.yaml_args <- list(
+  line.sep = eol(),
+  indent.mapping.sequence = TRUE,
+  handlers = list(logical = as_verbatim_lgl)
+)
+
+schemas_are_up_to_date <- function(old, new) {
+  old <- old[c("author", "affiliations")]
+  identical(old, new)
+}
+
+json_update <- function(old, new) {
+  new <- as_json(new)
+  if (schemas_are_up_to_date(old, new)) {
     return()
   }
-  list_assign(x, !!!json)
+  out <- if (is.null(old)) new else list_assign(old, !!!new)
+  list_drop_empty(out)
 }
 
 as_json <- function(x) {
@@ -23,28 +36,10 @@ separate_yaml_header <- function(x) {
   str_split_1(x, "(?m:^|\\R\\K)-{3}(?:\\R|$)")
 }
 
-as_verbatim_lgl <- function(x) {
-  x <- if_else(x, "true", "false")
-  structure(x, class = "verbatim")
-}
-
-get_eol <- function() {
-  if (.Platform$OS.type == "unix") "\n" else "\r\n" # nocov
-}
-
 yaml_inject <- function(x, lines) {
-  eol <- get_eol()
-  if (is_empty(x$affiliations)) {
-    x$affiliations <- NULL
-  }
-  yaml <- as.yaml(
-    x,
-    line.sep = eol,
-    indent.mapping.sequence = TRUE,
-    handlers = list(logical = as_verbatim_lgl)
-  )
+  yaml <- do.call(as.yaml, c(list(x), .yaml_args))
   out <- replace(lines, 2, yaml)
-  collapse(out, paste0("---", eol))
+  collapse(out, paste0("---", eol()))
 }
 
 has_yaml <- function(x) {
@@ -80,15 +75,28 @@ add_yaml_header <- function(x) {
   c("", "", x)
 }
 
-yaml_push <- function(what, file) {
+yaml_push <- function(x, file) {
+  UseMethod("yaml_push")
+}
+
+yaml_push.default <- function(x, file) {
+  old <- yaml::read_yaml(file)
+  json <- json_update(old, x)
+  if (is.null(json)) {
+    return(invisible())
+  }
+  do.call(yaml::write_yaml, c(list(json), file, .yaml_args))
+}
+
+yaml_push.qmd <- function(x, file) {
   text <- read_file(file)
   check_has_yaml(text)
   items <- separate_yaml_header(text)
   if (yaml_has_strippable(items[[2]])) {
     items <- add_yaml_header(items)
   }
-  json <- as_json(what)
-  json <- json_update(items[[2]], json)
+  old <- yaml.load(items[[2]])
+  json <- json_update(old, x)
   if (is.null(json)) {
     return(invisible())
   }